Deleting qsharp.json causes error squiggles to go haywire (#989)
When a qsharp.json is deleted from under a project, we find out about it
the next time `update_document` or `close_document` is called (i.e. the
next time we call `get_manifest` ). The document is put back into
single-file mode. This is all well and great, but we forget to remove
the `qsharp.json` project that the document _used_ to be in.

1. With this fix, in `update_document`, we check if the document changed
compilations, and if it did, we close the previous compilation if there
are no other open documents remaining for with it.

2. Also opportunistically refactoring some code in state.rs

3. Also adding a proper in-memory filesystem to the language service
unit tests so we can write more nuanced cases like this one.
 

Repro steps for the UX issue that this fixes:

- Have a project with one file:
```
project/qsharp.json
project/src/foo.qs
```

- Type some errors in `foo.qs` . 
- While `foo.qs` is still open in the editor and still has errors,
delete `qsharp.json`
- Type some more in `foo.qs`. Notice how the position of the squiggles
don't get updated as you type.

This also fixes #973!