Skip running memtest in merge groups (#999)
The memtest task was causing spurious failures during merge group builds
because it failed to write back a comment even though it ran
successfully (merge group builds do not associate back to a signle
issue/PR). This skips the comment step if `github.base_ref` is null, so
will still run but not try to write back specific results.