Remove shots from statful interpreter `run` API
This change removes the shots argument to the `run` API in stateful interpreter now that all callers pass 1. This avoids the need to wrap interpreter results in an additional vector and simplifies calling patterns. All consumers already passed 1 and had a shot loop around the call to `run` or utilized other APIs.

This also removes the need for a `reinit` function on the `Backend` trait, since now a simulator is only used once before being recreated.