Refactor common interpreter patterns
This change refactors the interpreter API to unify around some common patterns and reduce places that look like exceptions. Several of the APIs used what appeared to be different patterns for how they invoked the evaluator when it turns out they were all minor, surface differences except for a few differences in input values. Using a wrapper function clarifies where the common patterns are and makes it easier for future updates to reuse them.