Turn core/std lib into a crate (#1029)
This change makes core/std lib sources available via a crate instead of
loading through a macro in the compiler. General effect is that the
sources should only be compiled when changes are made to the library vs
the compiler speeding up workflows in the compiler.

This can also help in future use of the language services if needed to
have direct access to the library sources.

The tests have been updated to follow more established patterns for
tests used in the rest of the repository.