update pipeline to run sample projects; add multi file sample (#1040)
This PR updates our samples build stage in the pipeline to accept
projects. We detect if a sample is a project in `build.py` and execute
the correct `qsc` command accordingly.

Additionally, this PR adds a new command line flag to `qsc` and `qsi`:
`--qsharp-json`. This flag allows the user to specify a path to a
`qsharp.json` to compile a project via the command line. Previously, the
only way to compile a project via the command line was to `cd` into it.