More predictable init and better logging in VS Code tests (#1015)
This is an attempt at fixing or mitigating #963, but since I can't repro
that one reliably, we won't be sure for a while.

* Product change: fix VS Code log formatting.

Before:
![Screenshot 2024-01-12
110423](https://github.com/microsoft/qsharp/assets/16928427/9444ff05-c8b4-4edf-8614-8f7db7fe35f3)
After:
![Screenshot 2024-01-12
110110](https://github.com/microsoft/qsharp/assets/16928427/b83f2674-784f-498a-95b0-dc940c102f56)

Test changes: 

* Explicitly activate the Q# extension before the tests start. This
should eliminate issues where the activation itself is taking a long
time for whatever reason and timing out the test cases, which have a
reasonably conservative timeout of 2s.
* Write Q# logs to the console in the integration tests.
* Disable the ugly verbose vscode / webserver logs by default. They were
helpful to debug infrastructure issues but we don't need them at the
moment.