Skip macOS in PR builds, include integration tests (#1044)
This change updates the logic for the build matrix to skip macOS when
building for PRs and enables integration tests. This keeps the
contentions and slower macOS runners reserved for merge queue to main.

To make this work nicely with branch protection rules, a "Status Check"
task is introduced that depends on previous jobs, only executes on
failure of dependencies, and always fails. Branch protection treats this
job as required, and then PR build will require only those matrix
combinations enabled in that context while merge queues will require the
full matrix to run and pass.