Separate Debugger into a distinct struct from Interpreter (#1039)
The immediate, selfish motivation for this change is that I want to add
a parameter to the `Interpreter` constructor (in #1038) that is really
only relevant to debugger scenarios, and I didn't want to update every
single call to `Interpreter::new`.

But I think this is a good change regardless. Some methods and fields in
`Interpreter` are only applicable to debugging, and the rest are only
applicable to not-debugging. It's really two separate use cases and two
separate sets of data. I wanted to make this distinction clearer.

Very open to feedback and opinions, and suggestions for further cleanup.