[Missing API] Added BigIntAsBoolArray  (#1047)
There currently exist the following std lib functions in the
`Microsoft.Quantum.Convert` namespace:

 - `BoolArrayAsInt`
 - `IntAsBoolArray`
 - `BigIntAsBoolArray`

However, `BoolArrayAsBigInt` is missing (and it used to exist in the old
QDK too).
I think it makes sense, for completeness sake, to have it here as well,
so this PR introduces it.

I also added a test for `BigIntAsBoolArray` as it did not have an
associated test.