Consolidate samples, run notebooks in build
This change consolidates our samples under the top level samples folder by moving the three notebooks that were under pip/samples.

It also adds a new build step controlled by `--notebooks` that uses a virtual python environment to run the notebooks and verify the cells execute successfully. It does not verify the contents of the output, that can come in a later PR.