Return utf-16 line and column offsets directly from the language service and debugger (#1038)
Fixes #851

- Introduce the `Position` and `Range` types which contain line/column
offsets. These are the equivalent of `offset` and `Span` in the
compiler, but with line/columns instead of offsets.

- Given a source string and offset, we can generate line/column
positions in terms of both UTF-8 and UTF-16 code units. UTF-16 is useful
for JavaScript, UTF-8 everywhere else we might use them.

- The `BreakpointSpan` and `StackFrame`s returned by the debugger now
use these line/column `Range`s instead of offsets.

- All the language service methods now take `Position`s instead of
`offset`, and return `Range`s instead of `Span`s in any response
structs.

- DELETE all the code that has to deal with line/column/utf-16 mapping
in the JavaScript layer, getting rid of a lot of document loading and
reading code!