%%qsharp cell failures should mark the cell as failed (#1071)
When a `%%qsharp` cell throws a compiler or runtime error, the whole
should be marked as an error, just as when Python code throws an
exception.

Related: #1070 . When we test notebooks in automation, we want to catch
any failures in Q# cells.

VS Code before/after side by side:


![image](https://github.com/microsoft/qsharp/assets/16928427/b8938278-7535-4abf-be9b-b685e7865234)

Jupyter notebook before/after side by side:


![image](https://github.com/microsoft/qsharp/assets/16928427/2d669d0a-0b0f-43de-9bef-fa083dbe49a9)