Reverse Endianness (#1079)
Based on feedback, we are reversing the endianness of ket displays from
the QDK along with associated code. This includes reversing the display
for state dumps from within Q#, in the debugger, and in Jupyter output.
Included in this change is the revert of updates to kata content that
tried to accomodate the different ordering between exercise text and the
state display. The change also includes update to the new
`dump_operation` in Python so that it is consistent with the endianness
and outputs matrices that match the same conventions.

---------

Co-authored-by: Scott Carda <55811729+ScottCarda-MS@users.noreply.github.com>