Prefer open file contents to disk contents (#1110)
This seems like what we should be doing anyway, but has the benefit that
when using the language service or the debugger in a multi-file project,
then everything just works even if the files haven't been saved to disk
yet.

Putting up early for review, but in initial testing looked to behave as
expected.