Evaluator performance improvments (#1116)
This change improves performance for array literals by utilizing the
`.extend` function on the stack vector, which allows size hints to
factor into the memory allocation. It also separates the continuation
stack and the action stack such that the continuations can be much
smaller and easier to allocate. The change also introduces a new
benchmark for large array literals to verify changes effect on that
specific case, which is common for large resource estimation programs.

The change has minor (within noise) effect on small evaluations like
teleport and Deutch-Jozsa, while having between 5% and 7% improvement in
array append, update, and literal evaluation benchamarks. When testing
the df-chemistry sample, around a 9% performance improvement was
abserved with the XVIII-cas4-fb-64e-56o.fcidump file.