Move the Location struct to a more common place (#1106)
A `Location` refers to a span within a specific source file. It's
expressed as a source name and `Range`. See:
https://github.com/microsoft/vscode/blob/7923151da4b3567636c1346df9e8ba682744bbf3/src/vscode-dts/vscode.d.ts#L6581

It was defined in the language service protocol, but I'm moving it to a
more common location in `compiler/qsc` since I need the debugger to use
it in #1107.

To summarize this PR and #1038, here's how locations are represented in
the compiler vs. higher-level components that interface with VS Code
(debugger, language service):
 
| Compiler representation | Line/column representation |
|---|---|
| `u32` offset (into `SourceMap`) | `Position` (into specific source
file) |
| `Span` | `Range` |
| (`PackageId`, `Span`) | `Location` |