Split system and estimation (#1135)
This splits the `estimates` module inside `resource_estimator` into two
modules `estimates` and `system`. The `estimates` module contains the
generic physical resource estimation code that is based on traits. The
`system` implements these traits and handles system creation based on
JSON configuration.

The `system` module refers to types from the `estimates` module. But the
`estimates` trait does **_not_** refer to any types from the `system`
module.

By splitting the module, I also split the `Error` type such that each
module has it's own `Error` type, and the system module's `Error` type
can be created from the estimates `Error` type. Also the trait functions
have been changed to return a `String` error in order to allow for a
generic error that is not tight to any of the module's error type.