Support generating QIR with custom intrinsics (#1141)
This change adds support for generating QIR with custom intrinsic
callables. Any Q# operation or function declared as `body intrinsic`
that is not part of the recognized built-in callables will be inserted
as a call to an extern function in the QIR declared with the provided
arguments and with a `void` return (see limitations below). These
functions must be recognized and supported by the consumer of the QIR or
it may cause compilation or linking failure further down the pipeline.
Note that custom intrinsics will still fail simulation.

This fixes #825.

### Limitations
- A custom intrinsic that returns types other than `Unit` causes a
runtime error during code generation with help text that advises the use
of `Unit`:

![image](https://github.com/microsoft/qsharp/assets/10567287/b1c536e8-72f2-45b3-b17a-0775896ba429)

- Passing complex, memory backed types to custom intrinsics (`BigInt`,
arrays, tuples) will cause code generation to fail with a runtime error
mentioning the unsupported type:

![image](https://github.com/microsoft/qsharp/assets/10567287/908df8ac-7c86-4ad1-b90f-b5706a369bcc)

- While not a memory backed type, `Result` also cannot be used as an
argument or return because it would break logic used for deferring
measurements and potentially violate requirements of the Base Profile
spec (results should be write-only and tagged as such except in
measurements, for example). These cases are rejected with the same
runtime errors shown above.