Generate Docs API (#1150)
Adds a new module to the qsc for docs generation and moves the display
module from the language server to this new module. The docs generation
is used by the npm to build library API docs from the doc comments of
our Q# library code and puts those docs in the npm package. This is done
when the npm package is built.

---------

Co-authored-by: Bill Ticehurst <billti@microsoft.com>