Add DumpRegister
This change adds an implementation of `DumpRegister` as an intrinsic in the standard library with matching implementation in the evaluator. It works by getting a full state dump and calculating if the state corresponding to the given qubits is separable from the rest, returning an error if it is not.

Fixes #1138