Make FIR package store queries return `&Package` instead of `Option<&Package>` (#1183)
When we query a package from an FIR package store, we always expect the
package to be present in the store. Therefore, there is no need to
return an option (we always use `expect` to extract the package
reference from the option).

This change modifies the `get` and `get_mut` methods for
`fir::PackageStore` such that they return a reference to a package
instead of an option.