Generic code parameter in error correction trait (#1191)
The `ErrorCorrection` trait is generalized such that it uses a code
parameter instead of a code distance to compute the logical error rate,
number of physical qubits, and cycle time. The code parameter's type is
an associated type `Parameter` to the `ErrorCorrection` trait. It could
be a numeric type, e.g., to represent a code distance, but it could also
be a tuple type, to represent multiple parameters that influence the
metrics above. A new trait function is added to traverse the possible
space of values for the code parameter.

Further, this PR
- adds a test case to the system that checks consistency of the result
output, in order to detect changes that break this format,
- removes serialization from the `estimate` folder and moves it to the
`system` folder.