Context specific comparison of code parameters in RE (#1213)
In some applications, two code parameters cannot simply be compared
without taking into account the context of the error correction method
and the physical qubit. Therefore, instead of using `Ord` for
`ErrorCorrection::Parameter`, I introduce a trait function
`ErrorCorrection::code_parameter_cmp`, which compares two code
parameters in the context of the error correction itself and the
underlying physical qubit.

Further, I simplified the code and made some variable names and comments
consistent to not refer to T states in particular, when referring to
magic states.