QSC: display the path when emitting HIR/QIR fails (#1215)
I wasted a little time on this error as I could not figure out why the
emit was failing for me, so I hope the change makes sense to everyone.

When emitting HIR/QIR fails (e.g. directory permissions or output path
is invalid) it is not obvious where it failed because the code swallows
the path:

```
Error:   × could not emit QIR
  ╰─▶ Permission denied (os error 13)
```

or 

```
Error:   × could not emit QIR
  ╰─▶ No such file or directory (os error 2)
```

With this change the path is also shown:

```
Error:   × could not emit QIR file `./../qir/qir.ll`
  ╰─▶ No such file or directory (os error 2)
```

This is symmetrical to how the path is already shown for source file
errors

```
Error:   × could not read source file `../source/foo.qs`
  ╰─▶ No such file or directory (os error 2)
```