refactor compiler API in npm package (#1219)
This PR refactors the compiler API in the npm package to be extensible,
but backwards-compatible.
In doing so, this PR closes #1201.

This PR introduces `ProgramConfig`, an abstraction meant to encapsulate
whatever inputs go into
a compile. For now, this is just the source code and what language
features you want turned on.

We believe this will be more extensible going forward, as modifying a
config object type is easier to
make non-breaking than modifying a parameters directly.