Use workspace lint settings (#1224)
This change shifts from having common lints described at the top of each
crate's lib.rs to having them in a single top-level lints section of the
workspace Cargo.toml. This feature was new as of Rust 1.74, but we
hadn't switched to it ahead of the initial release.
Of note, a few crates were missing the top-level lints, so had lots of
fixup to do to get them up to the bar of the rest of code.