Alternative State Visualization  (#1162)
This change adds alternative state visualization to DumpMachine output
in the playground. The state is visualized as math formula (via LaTeX
created in Rust code in addition to a list of decimal amplitudes).
Recognition of **some** "nice" algebraic numbers (such as 2√3/5) and
exponents (such as 𝒆^(-𝝅·𝒊·2/3)) are supported. If the state is "ugly"
the generated LaTeX string is empty. Currently "ugly" states are states
with more than 16 basis vectors or more than two unrecognized decimal
numbers. Tests added.

For example:


![image](https://github.com/microsoft/qsharp/assets/60718360/747fceee-ad21-4da3-959e-73c158f523e1)

This resolves items 4 and 5 (partially) from the issue
https://github.com/microsoft/qsharp/issues/579
This resolves the old issue
https://github.com/microsoft/iqsharp/issues/446 from the legacy QDK.

Possible future work:
 * Recognize common normalization coefficient
 * Recognize separable states
 * Recognize well-known states such as: |0> |1> |+> |-> |Bell> |CATn>

Thanks to @minestarks, @billti, @swernli for help with propagating state
string from Rust code into playground display.

---------

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>
Co-authored-by: Mine Starks <mineyalc@microsoft.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>