Generic round based factory for estimation (#1230)
This extracts the modeling of generic round based factories to the
`estimates` module in order to be used in other systems. This round
based factory anticipates error correction models with a generic code
parameter. The exhaustive T-factory search is based on code distance and
remains part of the `system` module. Also, the specification of default
distillation units and modeling of custom distillation units via JSON is
part of the `system` module.