Fix generated type in replace qubit allocation pass (#1244)
During the replace qubit allocations pass, calls to qubit allocation
operations are generated, but the type of the call expression was always
set to `Qubit` regardless of whether the allocation was for a single
qubit or multiple qubits.

This change sets the appropriate types for the qubit allocation call
expression. `Qubit` for single qubit allocation and `Qubit[]` for
multi-qubit allocation.