Add integration tests for VS Code debugging, eliminate `FileAccessor` (#1107)
Depends on #1106

VSCode's debugging related APIs sometimes use filesystem paths instead
of URIs to represent sources. Previously, our solution was to have a
`FileAccessor` which tried to treat the path both ways (first as a
filesystem path, then a URI). But by being strategic about exactly where
we expect filesystem paths and sanitizing them, we can just consistently
use URIs in the rest of the code.

Additionally this PR introduces integration tests for the debugger.

We also fix two small-ish bugs here:
- launch.json variables like `{file}` and `{workspaceFolder}` weren't
treated properly on all platforms/filesystems
- when launching the debugger on an untitled file, the user may be
prompted to save the file. If they choose to save the file, the debugger
will fail to launch. (This is probably not as big an issue anymore,
since we disabled the auto-save behavior by default since I've opened
this PR.)