Make Base Profile result type check specific to entry expressions (#1256)
Fixes #1246

The current base profile result type check was overly broad, because it
ran the check on all fragments (not just entry expressions).

This was because during incremental compilation, entry expressions were
added to the AST package as a top-level statement. After that point, the
passes could not differentiate between the entry expression and a
top-level statement.

We don't _actually_ have to convert the entry expression to a statement
- we can just keep the entry expression in the `entry` field in the AST
and HIR. So, now, if passes want to apply any special treatment to the
entry expression, they have enough information to do so.