Connection string workspace auth (#1238)
This is the vast majority of the code for the connection string auth
work.

It's not as big as it looks due to a few things:

- I refactored a lot of long and complex functions into smaller
functions (and moved auth code to a new file)
- I ripped out the now dead CORS and bitcode proxy code.
- I fixed a couple drive-bys ('undefined' job name) and added some extra
error handling

I've still a little work to do around telemetry and error handling, and
being able to copy the connection string to the clipboard for use in a
Notebook, but it's 95% there for review now.