Use verbose flag when building wheels
This adds the `-v` flag to the pip wheel command in the build script which will help with future diagnosability, expecially in the release pipeline. With this addition, the underlying Rust build for the Python wheel is displayed to console/log so we can confirm which components are being built on which platforms.