Implement Runtime Capabilities Analysis (RCA) (#1164)
Runtime Capabilities Analysis (RCA) is the process of determining the
capabilities a quantum kernel needs to be able to run a particular
program. This implementation also identifies program elements that can
be pre-computed before execution on a quantum kernel and does not
consider these elements when determining the capabilities. Additionally,
this implementation provides details on why the program requires each
capability.

One way to look at it is that RCA answers the following questions:
- What capabilities are needed to run a program?
- Which program elements require these capabilities and why?
- Which program elements can be computed before execution in a quantum
kernel?

---------

Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>