Formatter Indentation and Delimiter Logic (#1289)
Added some state to the formatter for the purposes of better handling of
delimiters, i.e. `[]`, `()`, `{}`, and type parameter list delimiters
`<>`. The formatting of delimited blocks will be determined by whether
there is a newline in the whitespace following the open delimiter.
- If there is a newline, then the indentation is increased for the
length of the block, and separators, i.e. ',' and ';', require newlines
after them, with the closing delimiter having a newline in front of it.
- If there are no newlines, then delimiters have no spacing after the
opening and before the closing delimiters, except the brace delimiters
which have a singe space instead, and separators require a single space
after them.