Fixes Magic Getting Erased by Format in Notebooks (#1310)
This was caused by two different bugs that are addressed in this PR:
- When the input Q# uses `\r\n` line endings, the raw token lexer would
only consider the `\n` line ending, resulting in comment tokens with
trailing `\r` characters in them.
- In the notebooks, when there were newlines prior to the magic command,
the offsets were not mapped correctly.

In order to preserve the magic command in the notebooks from getting
removed as leading whitespace in a file (a formatter rule), the magic
command is replaced by a comment instead of whitespace.

Fixes #1308