RIR defer measurements pass
This change introduces an RIR -> RIR pass the defers measurements within a block. It preserves relative ordering of all instructions, and leaves the block in the format of:

 1. Operations
 2. Measurements
 3. Output recording
 4. Block terminator

For Base Profile programs where the whole program is a single block, this performs whole program measurement deferral. For programs with more complex control flow, this ensures that measurements appear at the end of the block in a way that could eventually support quantum logic separation.

This change also refactors some test utilities to support code reuse across tests for QIR codegen and passes.