Fix issues with circuit panel not refreshing properly (#1322)
Fixes issue where the Circuit panel will continue to display the
previous circuit, even after it's been updated with a new circuit that
fails to render.

Repro:
1. Generate circuit for a valid program, e.g. Entanglement.qs
2. Now generate circuit for a program that doesn't allocate any qubits,
e.g. Array.qs

![image](https://github.com/microsoft/qsharp/assets/16928427/299f68b1-d10d-4c4a-89cc-3486ed891eb1)


Also a small fix where we force the circuit panel to come to the
foreground when invoked through the code lens or at the beginning of a
debugging session.