Use dense FIR IDs across packages (#1296)
This change fixes three things:
- It reduces the memory footprint by using dense FIR IDs across
packages. This is accomplished by using a new lowerer instance for every
package such that IDs for blocks, statements, expressions and pats are
reset and start from zero.
- It fixes a bug where the debugger would not honor a breakpoint if the
statement that mapped to the breakpoint had the ID 0.
- As a consequence of changing the way the FIR lowerer is used, the
debugger would now randomly and inadvertently break into a core or
standard library statement if the statement ID happened to be the same
than the statement ID where a breakpoint is set in the user's code. This
problem is also addressed by this change.

This was done in collaboration with @swernli and @idavis.

---------

Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>