Skip `StmtId` entries in execution graph when not debugging
This change is an optimization that configures the lowerer to know when debugging is the intended scenario for the lowered FIR. When not debugging, no `StmtId` nodes are included in the execution graph, which skips the (minimal but non-zero) overhead of checking each statement for possible breakpoints. Only the debugger creates the interpreter and the underlying lowerer with these debug entries enabled, preserving behavior for those scenarios.