Mimalloc is vendored instead of fetched (#1320)
These changes copy a snapshot of the Mimalloc source from `v2.1.2` into
the mimalloc-sys folder so that we no longer need to use cmake to fetch
and prepare the source. This also fixes `aarch64` Windows
cross-compilation issues caused by cmake generators.

I have addressed all DevSkim feedback and nothing was applicable or
actionable.

I've run the release build and everything compiles. It shows a red X in
the checks as I cancelled the second half of the build to prevent the
release from going out.

Closes #1281