Fast inner-loop build watcher (#1324)
See the comments in dev.mjs

Basically after initial build with './build.py', simply run 'node
dev.mjs' and most changes to Rust, npm, VS Code, or playground can be
reloaded in seconds.

I expect this may have issues on Windows (e.g. maybe need to us
'wasm-pack.exe' there), so looking for feedback on if it works and time
it takes across platforms, as well as if this is something we should
check in.

Note that as part of this I cleaned up the npm imports to use the web
version EXCEPT where it's Node specific code. That's because this
watcher only rebuilds the web version for speed.

Also, I exported some functions from the build scripts so they could be
reused in this script.