RIR defer measurements and unreachable code passes (#1319)
This change introduces two RIR -> RIR passes: the first that defers
measurements within a block and a second that is a wrapper on a set of
unreachable code checks.

Deferring of measurements updates the RIR and preserves relative
ordering of all instructions, leaving the block in the format of:

 1. Operations
 2. Measurements and readout
 3. Output recording
 4. Block terminator

For Base Profile programs where the whole program is a single block,
this performs whole program measurement deferral. For programs with more
complex control flow, this ensures that measurements appear at the end
of the block in a way that could eventually support quantum logic
separation.

The unreachable code pass does not modify the RIR, instead acting as a
way to verify that RIR is well-formed and panicking on any unreachable
callables, blocks, or instructions.

This change also refactors some test utilities to support code reuse
across tests for QIR codegen and passes.

---------

Co-authored-by: César Zaragoza Cortés <cesarzc@microsoft.com>