Move RIR passes to qsc_rir
This change moves the RIR passes to be in the qsc_rir crate rather than qsc_codegen. This makes it easier for other parts of the code, notably partial evaluation, to use the passes and the builder utility without having to take a dependency on qsc_codegen. As part the change, the tests are updated to use the `Display` impl on RIR for test validation rather than depending on generating QIR snippets.