Move npm package (#1344)
As discussed, moving the npm package to a directory named after it.

Note that calling the directory name 'qsharp' for now in anticipation of
renaming the npm package from 'qsharp-lang' to 'qsharp' at some point in
the near future. That will be a bigger change that needs coordinating
with QCOM though, so will do separately.